refactor: extract GameService map cache and persistence part (#3878)
Move Hive load/save, auto-save mirror, and in-memory map cache helpers
into game_service_map_cache.dart; host game_service.dart keeps turn
resolution and new-game entry points only.

Refs #3878
