+++ SPEC/program/repo-and-packages.md
@@ -18,7 +18,7 @@ Monorepo layout:
 
 - **`core/services/app_event_handler/`** — shell-level bus wiring. **`AppEventHandler`** ([`app_event_handler.dart`](../../app/lib/core/services/app_event_handler/app_event_handler.dart)) dispatches `UIActionEvent` / `UISystemEvent` to Navigator/dialog APIs. Library parts: `app_event_handler_navigation.dart` (dialog, confirm, panel, pause, exit-to-shell navigation) and `app_event_handler_unit_panels.dart` (civilian/military/naval bottom-sheet hosts). **`AppEventHandlerScope`** ([`app_event_handler_scope.dart`](../../app/lib/core/services/app_event_handler/app_event_handler_scope.dart)) binds session-scoped bus listeners. Library parts: `app_event_handler_scope_train_orders.dart` (train-dialog order merge + combat-mode apply helper). Session subscriptions are split into `part` files by event family (observe, civilian work, naval, army, diplomacy, debug), each capped at 500 non-comment lines (`repo.app_event_handler_part_size`).
 - **`core/services/turn_resolution/`** — worker-isolate next-turn pipeline: [`TurnResolutionRunner`](../../app/lib/core/services/turn_resolution/turn_resolution_runner.dart) (library parts `turn_resolution_runner_types.dart`, `turn_resolution_runner_isolate_listener.dart`), [`turn_resolution_worker_isolate.dart`](../../app/lib/core/services/turn_resolution/turn_resolution_worker_isolate.dart), [`turn_resolution_result_codec.dart`](../../app/lib/core/services/turn_resolution/turn_resolution_result_codec.dart), [`TurnResolutionResultApplier`](../../app/lib/core/services/turn_resolution/turn_resolution_result_applier.dart), and [`turn_resolution_blocking_service.dart`](../../app/lib/core/services/turn_resolution/turn_resolution_blocking_service.dart) (clears `turnResolutionBlockingProvider` via root navigator context per `SPEC/program/app-ui-wiring.md`).
-- **`GameService`** (`app/lib/core/services/game_service/game_service.dart`) loads/saves games and advances turns. Library parts: `game_service_new_game_setup.dart` (create-new-game entry + bootstrap/finalize), `game_service_new_game_setup_pipeline.dart` (freeform/locked-full-init retry pipelines), `game_service_new_game_setup_maps.dart` (tile-map generation, warp links, persistence), `game_service_turn_resume.dart` (turn resolution, diplomacy resume APIs, bus persistence), and `game_service_turn_trace.dart` (merged trace export).
+- **`GameService`** (`app/lib/core/services/game_service/game_service.dart`) loads/saves games and advances turns. Library parts: `game_service_map_cache.dart` (Hive load/save, auto-save mirror, in-memory map cache), `game_service_new_game_setup.dart` (create-new-game entry + bootstrap/finalize), `game_service_new_game_setup_pipeline.dart` (freeform/locked-full-init retry pipelines), `game_service_new_game_setup_maps.dart` (tile-map generation, warp links, persistence), `game_service_turn_resume.dart` (turn resolution, diplomacy resume APIs, bus persistence), and `game_service_turn_trace.dart` (merged trace export).
 - **`core/services/debug/`** — ctdev/debug command handlers (`app_event_handler_debug_*.dart` library + parts), shared guard helpers (`debug_command_helpers.dart`), and [`DebugCommandSessionHandler`](../../app/lib/core/services/debug/debug_command_session_handler.dart) (applies debug mutations to the active session without importing feature panels).
 - **`core/services/ai/`** — blessed AI profile loading ([`blessed_ai_profile_loader.dart`](../../app/lib/core/services/ai/blessed_ai_profile_loader.dart)) and in-game profile resolution ([`ai_profile_resolution.dart`](../../app/lib/core/services/ai/ai_profile_resolution.dart)).
 - **`core/services/platform/`** — desktop shell startup ([`desktop_window_startup_service.dart`](../../app/lib/core/services/platform/desktop_window_startup_service.dart)).
