+++ app/lib/core/services/game_service/game_service.dart
@@ -10,38 +10,12 @@ import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_save/colonizethis_save.dart';
 import 'package:hive/hive.dart';
 
+part 'game_service_map_cache.dart';
 part 'game_service_new_game_setup.dart';
 part 'game_service_new_game_setup_pipeline.dart';
 part 'game_service_new_game_setup_maps.dart';
 part 'game_service_turn_resume.dart';
 part 'game_service_turn_trace.dart';
-
-/// Cached map data for a game (topology and tile maps for turn resolution).
-class _GameMapCache {
-  _GameMapCache({
-    required this.combinedTopology,
-    required this.tileMapByRegion,
-    required this.topologyByRegion,
-    this.warpLinks,
-  });
-  final MapTopology combinedTopology;
-  final Map<String, TileMapResult> tileMapByRegion;
-  final Map<String, MapTopology> topologyByRegion;
-  final List<WarpLink>? warpLinks;
-}
-
-/// Public record type for [GameService.getMapData] (Refs #2575 Phase 4).
-/// Lets callers replace `dynamic` with an explicit type while still using
-/// record-style access (`mapData.combinedTopology`, etc.).
-typedef GameMapData = ({
-  MapTopology combinedTopology,
-  Map<String, TileMapResult> tileMapByRegion,
-  Map<String, MapTopology> topologyByRegion,
-  List<WarpLink>? warpLinks,
-});
-
-/// Pass milestones for in-app tile map generation (SPEC/program/logging/map-generation.md).
-final _mapGenPassLog = packageLogger('tile_map');
 
 /// Loads/saves games and advances turn. SPEC/project/phase-1: app invokes TurnResolver and persists via colonizethis_save.
 /// Phase 2: createNewGame uses full game-setup pipeline; nextTurn requires cached/persisted map data.
@@ -80,146 +54,26 @@ class GameService {
   /// Populated when creating a new game or when loading a game with persisted map data.
   final Map<String, _GameMapCache> _mapCache = {};
 
-  _GameMapCache _requireMapData(String gameId) {
-    final cached = _mapCache[gameId];
-    if (cached != null) return cached;
-    final mapData = _adapter.loadMapData(_box, gameId);
-    final loaded = _GameMapCache(
-      combinedTopology: mapData.combinedTopology,
-      tileMapByRegion: mapData.tileMapByRegion,
-      topologyByRegion: mapData.topologyByRegion,
-      warpLinks: mapData.warpLinks,
-    );
-    _mapCache[gameId] = loaded;
-    return loaded;
-  }
-
-  /// Loads game by id. Returns null if not found or required map data is missing/invalid.
-  /// Populates _mapCache on successful load so map rendering and turn resolution work.
-  Game? loadGame(String gameId) {
-    final game = _adapter.load(_box, gameId);
-    if (game == null) return null;
-    try {
-      _requireMapData(gameId);
-      return game;
-    } catch (e, st) {
-      packageLogger().e(
-        'required map data missing/invalid for gameId=$gameId',
-        error: e,
-        stackTrace: st,
-      );
-      return null;
-    }
-  }
-
-  /// Returns map data for [gameId] from cache or storage.
-  ///
-  /// When [_mapCache] already holds [gameId], returns immediately without reading game
-  /// JSON from Hive (avoids redundant adapter load on UI hot paths). Otherwise loads
-  /// game JSON to verify existence, then ensures map data is loaded.
-  ///
-  /// Returns null only when no game exists for [gameId]. For existing games, map data
-  /// is required and missing/invalid map data raises [StateError].
-  GameMapData? getMapData(String gameId) {
-    final cached = _mapCache[gameId];
-    if (cached != null) {
-      return (
-        combinedTopology: cached.combinedTopology,
-        tileMapByRegion: cached.tileMapByRegion,
-        topologyByRegion: cached.topologyByRegion,
-        warpLinks: cached.warpLinks,
-      );
-    }
-    final gameExists = _adapter.load(_box, gameId) != null;
-    if (!gameExists) return null;
-    final cache = _requireMapData(gameId);
-    return (
-      combinedTopology: cache.combinedTopology,
-      tileMapByRegion: cache.tileMapByRegion,
-      topologyByRegion: cache.topologyByRegion,
-      warpLinks: cache.warpLinks,
-    );
-  }
-
-  ({
-    MapTopology combinedTopology,
-    Map<String, TileMapResult> tileMapByRegion,
-    Map<String, MapTopology> topologyByRegion,
-    List<WarpLink>? warpLinks,
-  })
-  _requiredMapDataView(String gameId) {
-    final cache = _requireMapData(gameId);
-    return (
-      combinedTopology: cache.combinedTopology,
-      tileMapByRegion: cache.tileMapByRegion,
-      topologyByRegion: cache.topologyByRegion,
-      warpLinks: cache.warpLinks,
-    );
-  }
-
   /// Optional strip for session-only observe control overrides before persist.
   Game Function(Game)? prepareGameForPersistence;
 
+  /// Loads game by id. Returns null if not found or required map data is missing/invalid.
+  Game? loadGame(String gameId) => _gameServiceLoadGame(this, gameId);
+
+  /// Returns map data for [gameId] from cache or storage.
+  GameMapData? getMapData(String gameId) => _gameServiceGetMapData(this, gameId);
+
   /// Saves game to storage.
-  void saveGame(Game game) {
-    final toSave = prepareGameForPersistence?.call(game) ?? game;
-    _adapter.save(_box, toSave);
-  }
+  void saveGame(Game game) => _gameServiceSaveGame(this, game);
 
   /// Lists all saved game ids.
-  List<String> listGameIds() => _adapter.listGameIds(_box);
+  List<String> listGameIds() => _gameServiceListGameIds(this);
 
-  /// Whether the Hive auto-save slot is playable. Clears invalid slots. SPEC/program/save-load.md.
-  bool hasValidAutoSave() => _adapter.hasValidAutoSave(_box);
+  /// Whether the Hive auto-save slot is playable.
+  bool hasValidAutoSave() => _gameServiceHasValidAutoSave(this);
 
-  /// Loads the auto-save slot into memory cache under [Game.id]. Returns null if missing/invalid.
-  Game? loadAutoSaveGame() {
-    if (!_adapter.hasValidAutoSave(_box)) {
-      return null;
-    }
-    final game = _adapter.load(_box, kAutoSaveSlotId);
-    if (game == null) {
-      return null;
-    }
-    try {
-      final mapData = _adapter.loadMapData(_box, kAutoSaveSlotId);
-      _mapCache[game.id] = _GameMapCache(
-        combinedTopology: mapData.combinedTopology,
-        tileMapByRegion: mapData.tileMapByRegion,
-        topologyByRegion: mapData.topologyByRegion,
-        warpLinks: mapData.warpLinks,
-      );
-      return game;
-    } catch (e, st) {
-      packageLogger().e(
-        'save: loadAutoSaveGame failed',
-        error: e,
-        stackTrace: st,
-      );
-      _adapter.delete(_box, kAutoSaveSlotId);
-      return null;
-    }
-  }
-
-  void _mirrorAutoSave(Game game) {
-    try {
-      final md = _requiredMapDataView(game.id);
-      _adapter.saveAutoSave(
-        _box,
-        game,
-        tileMapByRegion: md.tileMapByRegion,
-        topologyByRegion: md.topologyByRegion,
-        combinedTopology: md.combinedTopology,
-        warpLinks: md.warpLinks,
-      );
-    } catch (e, st) {
-      packageLogger().e(
-        'save: auto-save mirror failed',
-        error: e,
-        stackTrace: st,
-      );
-    }
-  }
+  /// Loads the auto-save slot into memory cache under [Game.id].
+  Game? loadAutoSaveGame() => _gameServiceLoadAutoSaveGame(this);
 
   /// Resolves one turn. Returns [TurnResolutionComplete] with new game (and persists),
   /// or a pending result: [TurnResolutionPendingOvertures], [TurnResolutionPendingIntervention],
