+++ app/lib/core/services/game_service/game_service_new_game_setup_maps.dart
@@ -96,6 +96,6 @@ void _gameServicePersistNewGame(
     warpLinks: result.warpLinks,
   );
   service.saveGame(result.game);
-  service._mirrorAutoSave(result.game);
+  _gameServiceMirrorAutoSave(service, result.game);
   service.eventBus?.emit(NewGameCreatedEvent(gameId: result.game.id));
 }
